Added iventory overview in UI
