+++ Assets/Prefabs/UI.prefab
@@ -1658,6 +1658,10 @@ PrefabInstance:
       propertyPath: m_Name
       value: Pause Menu
       objectReference: {fileID: 0}
+    - target: {fileID: 878275491582344413, guid: 9fe9f288f7c235b4e9734d39322ef273, type: 3}
+      propertyPath: m_IsActive
+      value: 1
+      objectReference: {fileID: 0}
     - target: {fileID: 2231681413517239919, guid: 9fe9f288f7c235b4e9734d39322ef273, type: 3}
       propertyPath: m_Pivot.x
       value: 0.5
@@ -1769,6 +1773,22 @@ PrefabInstance:
     - target: {fileID: 2431296962032681393, guid: 9fe9f288f7c235b4e9734d39322ef273, type: 3}
       propertyPath: m_OnClick.m_PersistentCalls.m_Calls.Array.data[0].m_Arguments.m_ObjectArgumentAssemblyTypeName
       value: UnityEngine.Object, UnityEngine
+      objectReference: {fileID: 0}
+    - target: {fileID: 4788097791706332291, guid: 9fe9f288f7c235b4e9734d39322ef273, type: 3}
+      propertyPath: m_AnchorMax.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 4788097791706332291, guid: 9fe9f288f7c235b4e9734d39322ef273, type: 3}
+      propertyPath: m_AnchorMin.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 4788097791706332291, guid: 9fe9f288f7c235b4e9734d39322ef273, type: 3}
+      propertyPath: m_AnchoredPosition.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 4788097791706332291, guid: 9fe9f288f7c235b4e9734d39322ef273, type: 3}
+      propertyPath: m_AnchoredPosition.y
+      value: 0
       objectReference: {fileID: 0}
     - target: {fileID: 7941131242839734252, guid: 9fe9f288f7c235b4e9734d39322ef273, type: 3}
       propertyPath: m_OnClick.m_PersistentCalls.m_Calls.Array.data[0].m_Target
