+++ Assets/Scripts/UI/UIInventory.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: dc6df6d053c130343b31cbad1b50568c
+guid: 31495fe8539a50b4b9052f15cbd4214b
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
