+++ Assets/Sprites/Items.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 876093bcaab49f9498c49885a62f31fe
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
